+++ packages/reward-root-submitter/reward_root_submitter/main.py
@@ -6,7 +6,9 @@ import pyarrow.parquet as pq
 from hexbytes import HexBytes
 from web3 import Web3
 
+from .config import Config
 from .contracts import RewardPool
+from .utils import get_roots_s3, get_roots_subgraph
 
 NULL_HEX = HexBytes(
     "0x0000000000000000000000000000000000000000000000000000000000000000"
@@ -15,6 +17,8 @@ NULL_HEX = HexBytes(
 EMPTY_MARKER_HEX = HexBytes(
     "0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF"
 )
+
+DEFAULT_MAX_PAST_BLOCKS = 34560  # 2 days (1 block every 5s)
 
 
 @dataclass
@@ -94,4 +98,26 @@ def process_file(reward_output_filename, config):
         merkle_root_details.merkle_root_hash,
         config.reward_root_submitter_address,
         config.reward_root_submitter_private_key,
+    )
+
+
+def get_all_unsubmitted_roots(config: Config):
+    evm_node = config.evm_full_node_url
+    w3 = Web3(Web3.HTTPProvider(evm_node))
+    current_block = w3.eth.get_block("latest")["number"]
+    min_scan_block = current_block - DEFAULT_MAX_PAST_BLOCKS
+
+    s3_df = get_roots_s3(config, min_scan_block)
+    subgraph_df = get_roots_subgraph(config, min_scan_block)
+    left_exclude_join_df = s3_df.merge(
+        subgraph_df,
+        how="left",
+        on=["reward_program_id", "payment_cycle"],
+        indicator=True,
+    ).copy()
+    missing_roots_df = left_exclude_join_df[
+        left_exclude_join_df["_merge"] == "left_only"
+    ].drop("_merge", axis=1)
+    logging.info(
+        f"Total of {len(missing_roots_df)} out of {len(s3_df)} roots are missing "
     )
